Karabiner-elements instant 'caps_lock → escape'
It seems like this config enables _instant_ 'caps_lock → escape'

In the past, I had to wait 5-10ms after pressing caps_lock but now it
seems like it is instant.

This might not seem like a big deal, but it is for Vim users because we
press escape all the time to switch between normal, visual, and insert
mode.